requirements.txt
streamlit>=1.39
supabase>=2.9
python-dotenv>=1.0
